Add files via upload
This is just a simple script that asks what IP address you would like to scan.  It then scans with nmap based on which type of scan you choose, then asks you what kind of gobuster scan you'd like to run.  Finally it runs dirb and spits them all out into a .txt file labeled with the scanned IP address and an identifier on the scan type.